chore: Update gopkg.in/yaml.v3 dependency
Remove direct dependency and add as indirect dependency

This change addresses a potential conflict and simplifies the
dependency management.  The `gopkg.in/yaml.v3` package was
previously listed as both a direct and indirect dependency. This
commit removes the direct dependency, leaving only the indirect
one, improving clarity and reducing potential issues.